feat: Add image generation to all providers that support it (#357)
* feat: add transcription to all providers that support it

* feat: gemini streaming

* chore: fix clippy warnings

* feat: together streaming

* refactor: move transcription preable to constant

* chore: fix clippy

* fix: enable blob by default

* fix: lock methods and imports behind feature flag

* chore: fmt

* feat: add huggingface transcription, remove feature flag

* feat: streaming for most openai type providers

* feat: add streaming to remaining providers

* chore: formatting

* fix: gemini streaming wasn't added to mod

* chore: unused imports in example

* chore: unused import

* chore: run fmt

* feat: openai & azure image generation

* chore: reorganize a bit

* refactor: simplify oai-compatible streaming

* feat: huggingface image generation

* chore: run fmt & clippy

* feat: hyperbolic image generation

* refactor: break up openai

* fix: lock changes behind feature flag

* chore: remove unused import

* refactor: change size to width/height, update examples

* fix: error with examples

* chore: remove unused imports

* fix: feature in lib not gated behind flag

* fix: used wrong flag

* fix: import not behind flag

* fix: more flag issues

---------
