+++ rig-core/src/lib.rs
@@ -85,6 +85,8 @@ pub mod cli_chatbot;
 pub mod completion;
 pub mod embeddings;
 pub mod extractor;
+#[cfg(feature = "image")]
+pub mod image_generation;
 pub(crate) mod json_utils;
 pub mod loaders;
 pub mod one_or_many;
