+++ rig-core/src/providers/huggingface/client.rs
@@ -2,7 +2,12 @@ use std::fmt::Display;
 
 use super::completion::CompletionModel;
 use crate::agent::AgentBuilder;
+#[cfg(feature = "image")]
+use crate::image_generation::ImageGenerationError;
+#[cfg(feature = "image")]
+use crate::providers::huggingface::image_generation::ImageGenerationModel;
 use crate::providers::huggingface::transcription::TranscriptionModel;
+use crate::transcription::TranscriptionError;
 
 // ================================================================
 // Main Huggingface Client
@@ -36,10 +41,27 @@ impl SubProvider {
     /// Get the transcription endpoint for the SubProvider
     /// Required because Huggingface Inference requires the model
     /// in the url and in the request body.
-    pub fn transcription_endpoint(&self, model: &str) -> String {
+    pub fn transcription_endpoint(&self, model: &str) -> Result<String, TranscriptionError> {
         match self {
-            SubProvider::HFInference => format!("hf-inference/models/{}", model),
-            _ => panic!("transcription endpoint is not supported yet for {}", self),
+            SubProvider::HFInference => Ok(format!("/{}", model)),
+            _ => Err(TranscriptionError::ProviderError(format!(
+                "transcription endpoint is not supported yet for {}",
+                self
+            ))),
+        }
+    }
+
+    /// Get the image generation endpoint for the SubProvider
+    /// Required because Huggingface Inference requires the model
+    /// in the url and in the request body.
+    #[cfg(feature = "image")]
+    pub fn image_generation_endpoint(&self, model: &str) -> Result<String, ImageGenerationError> {
+        match self {
+            SubProvider::HFInference => Ok(format!("/{}", model)),
+            _ => Err(ImageGenerationError::ProviderError(format!(
+                "image generation endpoint is not supported yet for {}",
+                self
+            ))),
         }
     }
 
@@ -124,7 +146,9 @@ pub struct Client {
 impl Client {
     /// Create a new Huggingface client with the given API key.
     pub fn new(api_key: &str) -> Self {
-        Self::from_url(api_key, HUGGINGFACE_API_BASE_URL, SubProvider::HFInference)
+        let base_url =
+            format!("{}/{}", HUGGINGFACE_API_BASE_URL, SubProvider::HFInference).replace("//", "/");
+        Self::from_url(api_key, &base_url, SubProvider::HFInference)
     }
 
     /// Create a new Client with the given API key and base API URL.
@@ -193,8 +217,25 @@ impl Client {
     ///
     /// let completion_model = client.transcription_model(huggingface::WHISPER_LARGE_V3);
     /// ```
+    ///
     pub fn transcription_model(&self, model: &str) -> TranscriptionModel {
         TranscriptionModel::new(self.clone(), model)
+    }
+
+    /// Create a new image generation model with the given name
+    ///
+    /// # Example
+    /// ```
+    /// use rig::providers::huggingface::{Client, self}
+    ///
+    /// // Initialize the Huggingface client
+    /// let client = Client::new("your-huggingface-api-key");
+    ///
+    /// let completion_model = client.image_generation_model(huggingface::WHISPER_LARGE_V3);
+    /// ```
+    #[cfg(feature = "image")]
+    pub fn image_generation_model(&self, model: &str) -> ImageGenerationModel {
+        ImageGenerationModel::new(self.clone(), model)
     }
 
     /// Create an agent builder with the given completion model.
