+++ rig-core/src/providers/huggingface/mod.rs
@@ -12,6 +12,9 @@
 
 pub mod client;
 pub mod completion;
+
+#[cfg(feature = "image")]
+pub mod image_generation;
 pub mod streaming;
 pub mod transcription;
 
@@ -20,4 +23,7 @@ pub use completion::{
     GEMMA_2, META_LLAMA_3_1, PHI_4, QWEN2_5, QWEN2_5_CODER, QWEN2_VL, QWEN_QVQ_PREVIEW,
     SMALLTHINKER_PREVIEW,
 };
+
+#[cfg(feature = "image")]
+pub use image_generation::{FLUX_1, KOLORS, STABLE_DIFFUSION_3};
 pub use transcription::{WHISPER_LARGE_V3, WHISPER_LARGE_V3_TURBO, WHISPER_SMALL};
